Update and rename TP_Ant_10101.html to main.html
Reflects changes made to the .js file. Also changed the title of the browser tab to include CSS since CSS is used here and deserves a spot there in the browser tab along with the other languages (and p5). Also added more to some comments. Hopefully that will help newcomers a bit.